refactor: 액션 폴더 분리
